daemon: remove duplicated code in both states
